+++ libtiff/tif_dirwrite.c
@@ -1194,7 +1194,7 @@ static int TIFFWriteDirectorySec(TIFF *tif, int isimage, int imagedone,
             tif->tif_dir.td_dirdatasize_write += 8 + ndir * 20 + 8;
 
         /* Setup a new directory within first pass. */
-        dir = (TIFFDirEntry *)_TIFFmallocExt(tif, (tmsize_t)(ndir * sizeof(TIFFDirEntry)));
+        dir = (TIFFDirEntry *)_TIFFmallocExt(tif, (tmsize_t)((size_t)ndir * sizeof(TIFFDirEntry)));
         if (dir == NULL)
         {
             TIFFErrorExtR(tif, module, "Out of memory");
@@ -1530,7 +1530,7 @@ static int TIFFWriteDirectoryTagSampleformatArray(TIFF *tif, uint32_t *ndir,
     void *conv;
     uint32_t i;
     int ok;
-    conv = _TIFFmallocExt(tif, (tmsize_t)(count * sizeof(double)));
+    conv = _TIFFmallocExt(tif, (tmsize_t)((size_t)count * sizeof(double)));
     if (conv == NULL)
     {
         TIFFErrorExtR(tif, module, "Out of memory");
@@ -1669,7 +1669,7 @@ static int TIFFWriteDirectoryTagShortPerSample(TIFF *tif, uint32_t *ndir,
         return (TIFFWriteDirectoryTagCheckedShortArray(
             tif, ndir, dir, tag, tif->tif_dir.td_samplesperpixel, NULL));
     }
-    m = (uint16_t *)_TIFFmallocExt(tif, (tmsize_t)(tif->tif_dir.td_samplesperpixel * sizeof(uint16_t)));
+    m = (uint16_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)tif->tif_dir.td_samplesperpixel * sizeof(uint16_t)));
     if (m == NULL)
     {
         TIFFErrorExtR(tif, module, "Out of memory");
@@ -1739,7 +1739,7 @@ static int TIFFWriteDirectoryTagLong8Array(TIFF *tif, uint32_t *ndir,
     ** For classic tiff we want to verify everything is in range for long
     ** and convert to long format.
     */
-    p = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)(count * sizeof(uint32_t)));
+    p = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)count * sizeof(uint32_t)));
     if (p == NULL)
     {
         TIFFErrorExtR(tif, module, "Out of memory");
@@ -1792,7 +1792,7 @@ static int TIFFWriteDirectoryTagSlong8Array(TIFF *tif, uint32_t *ndir,
     ** For classic tiff we want to verify everything is in range for signed-long
     ** and convert to signed-long format.
     */
-    p = (int32_t *)_TIFFmallocExt(tif, (tmsize_t)(count * sizeof(uint32_t)));
+    p = (int32_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)count * sizeof(uint32_t)));
     if (p == NULL)
     {
         TIFFErrorExtR(tif, module, "Out of memory");
@@ -2011,7 +2011,7 @@ static int TIFFWriteDirectoryTagLongLong8Array(TIFF *tif, uint32_t *ndir,
         ** and convert to long format.
         */
 
-        uint32_t *p = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)(count * sizeof(uint32_t)));
+        uint32_t *p = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)count * sizeof(uint32_t)));
         uint32_t *q;
         uint64_t *ma;
         uint32_t mb;
@@ -2041,7 +2041,7 @@ static int TIFFWriteDirectoryTagLongLong8Array(TIFF *tif, uint32_t *ndir,
     }
     else
     {
-        uint16_t *p = (uint16_t *)_TIFFmallocExt(tif, (tmsize_t)(count * sizeof(uint16_t)));
+        uint16_t *p = (uint16_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)count * sizeof(uint16_t)));
         uint16_t *q;
         uint64_t *ma;
         uint32_t mb;
@@ -2101,7 +2101,7 @@ static int TIFFWriteDirectoryTagIfdIfd8Array(TIFF *tif, uint32_t *ndir,
     ** and convert to long format.
     */
 
-    p = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)(count * sizeof(uint32_t)));
+    p = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)count * sizeof(uint32_t)));
     if (p == NULL)
     {
         TIFFErrorExtR(tif, module, "Out of memory");
@@ -2161,15 +2161,15 @@ static int TIFFWriteDirectoryTagColormap(TIFF *tif, uint32_t *ndir,
         return 1;
     }
 
-    n = (uint16_t *)_TIFFmallocExt(tif, (tmsize_t)(3U * m * sizeof(uint16_t)));
+    n = (uint16_t *)_TIFFmallocExt(tif, (tmsize_t)(3U * (size_t)m * sizeof(uint16_t)));
     if (n == NULL)
     {
         TIFFErrorExtR(tif, module, "Out of memory");
         return (0);
     }
-    _TIFFmemcpy(&n[0], tif->tif_dir.td_colormap[0], (tmsize_t)(m * sizeof(uint16_t)));
-    _TIFFmemcpy(&n[m], tif->tif_dir.td_colormap[1], (tmsize_t)(m * sizeof(uint16_t)));
-    _TIFFmemcpy(&n[2 * m], tif->tif_dir.td_colormap[2], (tmsize_t)(m * sizeof(uint16_t)));
+    _TIFFmemcpy(&n[0], tif->tif_dir.td_colormap[0], (tmsize_t)((size_t)m * sizeof(uint16_t)));
+    _TIFFmemcpy(&n[m], tif->tif_dir.td_colormap[1], (tmsize_t)((size_t)m * sizeof(uint16_t)));
+    _TIFFmemcpy(&n[2 * m], tif->tif_dir.td_colormap[2], (tmsize_t)((size_t)m * sizeof(uint16_t)));
     o = TIFFWriteDirectoryTagCheckedShortArray(tif, ndir, dir, TIFFTAG_COLORMAP,
                                                3 * m, n);
     _TIFFfreeExt(tif, n);
@@ -2213,10 +2213,10 @@ static int TIFFWriteDirectoryTagTransferfunction(TIFF *tif, uint32_t *ndir,
     {
         if (!_TIFFmemcmp(tif->tif_dir.td_transferfunction[0],
                          tif->tif_dir.td_transferfunction[2],
-                         (tmsize_t)(m * sizeof(uint16_t))) &&
+                         (tmsize_t)((size_t)m * sizeof(uint16_t))) &&
             !_TIFFmemcmp(tif->tif_dir.td_transferfunction[0],
                          tif->tif_dir.td_transferfunction[1],
-                         (tmsize_t)(m * sizeof(uint16_t))))
+                         (tmsize_t)((size_t)m * sizeof(uint16_t))))
             n = 1;
     }
     if (dir == NULL) /* Just evaluate IFD data size and increment ndir. */
@@ -2225,20 +2225,20 @@ static int TIFFWriteDirectoryTagTransferfunction(TIFF *tif, uint32_t *ndir,
         return 1;
     }
 
-    o = (uint16_t *)_TIFFmallocExt(tif, (tmsize_t)(n * m * sizeof(uint16_t)));
+    o = (uint16_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)n * m * sizeof(uint16_t)));
     if (o == NULL)
     {
         TIFFErrorExtR(tif, module, "Out of memory");
         return (0);
     }
     _TIFFmemcpy(&o[0], tif->tif_dir.td_transferfunction[0],
-                (tmsize_t)(m * sizeof(uint16_t)));
+                (tmsize_t)((size_t)m * sizeof(uint16_t)));
     if (n > 1)
         _TIFFmemcpy(&o[m], tif->tif_dir.td_transferfunction[1],
-                    (tmsize_t)(m * sizeof(uint16_t)));
+                    (tmsize_t)((size_t)m * sizeof(uint16_t)));
     if (n > 2)
         _TIFFmemcpy(&o[2 * m], tif->tif_dir.td_transferfunction[2],
-                    (tmsize_t)(m * sizeof(uint16_t)));
+                    (tmsize_t)((size_t)m * sizeof(uint16_t)));
     p = TIFFWriteDirectoryTagCheckedShortArray(
         tif, ndir, dir, TIFFTAG_TRANSFERFUNCTION, n * m, o);
     _TIFFfreeExt(tif, o);
@@ -2260,7 +2260,7 @@ static int TIFFWriteDirectoryTagSubifd(TIFF *tif, uint32_t *ndir,
         uint64_t *pa;
         uint32_t *pb;
         uint16_t p;
-        o = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)(tif->tif_dir.td_nsubifd * sizeof(uint32_t)));
+        o = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)tif->tif_dir.td_nsubifd * sizeof(uint32_t)));
         if (o == NULL)
         {
             TIFFErrorExtR(tif, module, "Out of memory");
@@ -2588,7 +2588,7 @@ static int TIFFWriteDirectoryTagCheckedRationalArray(TIFF *tif, uint32_t *ndir,
         EvaluateIFDdatasizeWrite(tif, count * 2, sizeof(uint32_t), ndir);
         return 1;
     }
-    m = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)(count * 2 * sizeof(uint32_t)));
+    m = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)count * 2 * sizeof(uint32_t)));
     if (m == NULL)
     {
         TIFFErrorExtR(tif, module, "Out of memory");
@@ -2624,7 +2624,7 @@ static int TIFFWriteDirectoryTagCheckedSrationalArray(TIFF *tif, uint32_t *ndir,
         EvaluateIFDdatasizeWrite(tif, count * 2, sizeof(int32_t), ndir);
         return 1;
     }
-    m = (int32_t *)_TIFFmallocExt(tif, (tmsize_t)(count * 2 * sizeof(int32_t)));
+    m = (int32_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)count * 2 * sizeof(int32_t)));
     if (m == NULL)
     {
         TIFFErrorExtR(tif, module, "Out of memory");
@@ -2661,7 +2661,7 @@ TIFFWriteDirectoryTagCheckedRationalDoubleArray(TIFF *tif, uint32_t *ndir,
         EvaluateIFDdatasizeWrite(tif, count * 2, sizeof(uint32_t), ndir);
         return 1;
     }
-    m = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)(count * 2 * sizeof(uint32_t)));
+    m = (uint32_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)count * 2 * sizeof(uint32_t)));
     if (m == NULL)
     {
         TIFFErrorExtR(tif, module, "Out of memory");
@@ -2696,7 +2696,7 @@ static int TIFFWriteDirectoryTagCheckedSrationalDoubleArray(
         EvaluateIFDdatasizeWrite(tif, count * 2, sizeof(int32_t), ndir);
         return 1;
     }
-    m = (int32_t *)_TIFFmallocExt(tif, (tmsize_t)(count * 2 * sizeof(int32_t)));
+    m = (int32_t *)_TIFFmallocExt(tif, (tmsize_t)((size_t)count * 2 * sizeof(int32_t)));
     if (m == NULL)
     {
         TIFFErrorExtR(tif, module, "Out of memory");
